GetDebtors, getOverdueDocuments and add returnDate to Booking

+++ src/User.java
@@ -1,3 +1,4 @@
+import javax.xml.crypto.Data;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.sql.Statement;
@@ -18,7 +19,7 @@ public abstract class User {
         if (Database.isLibrarian(idLibrarian)) {
             PreparedStatement preparedStatement;
             try {
-                preparedStatement = Database.connection.prepareStatement("UPDATE users name = ?, phoneNumber = ?, address = ?, debt = ?, isFacultyMember = ?, password = ?, isLibrarian = ? WHERE id = ?");
+                preparedStatement = Database.connection.prepareStatement("UPDATE users SET name = ?, phoneNumber = ?, address = ?, debt = ?, isFacultyMember = ?, password = ?, isLibrarian = ? WHERE id = ?");
                 preparedStatement.setString(1, name);
                 preparedStatement.setString(2, phoneNumber);
                 preparedStatement.setString(3, address);
@@ -36,18 +37,26 @@ public abstract class User {
         }
     }
 
-    public void DeleteUserDB(int idLibrarian){
+    public int DeleteUserDB(int idLibrarian){
         if(Database.isLibrarian(idLibrarian)) {
-            try {
-                PreparedStatement ps = Database.connection.prepareStatement("delete from users where id = ?");
-                ps.setInt(1, this.id);
-                ps.executeUpdate();
-            }catch (Exception e){
-                System.out.println("Error in DeleteUSERdb " + e.toString());
+            Database db = new Database();
+            if(db.getUserDocuments((Patron) this).isEmpty()) {
+                try {
+                    PreparedStatement ps = Database.connection.prepareStatement("delete from users where id = ?");
+                    ps.setInt(1, this.id);
+                    ps.executeUpdate();
+                    return 0;
+                } catch (Exception e) {
+                    System.out.println("Error in DeleteUSERdb " + e.toString());
+                }
+            }
+            else{
+                return -1;
             }
         }
         else {
             System.out.println("Error: User does not have access to delete user");
         }
+        return -1;
     }
 }
